Backend/requirements.txt
asgiref==3.9.1
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
dj-database-url==3.0.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.20.0
flashtext==2.7
fsspec==2025.9.0
gunicorn==21.2.0
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pytesseract==0.3.10
python-dotenv==1.1.1
PyYAML==6.0.3
rapidfuzz==3.9.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.7.0
redis==5.2.0
reportlab==4.2.5
